fix: prevent using mocks in rpc
