+++ packages/vitest/src/runtime/rpc.ts
@@ -1,4 +1,26 @@
-import { getWorkerState, withSafeTimers } from '../utils'
+import { getWorkerState } from '../utils'
+import {
+  setTimeout as safeSetTimeout,
+} from '../utils/timers'
+
+const safeRandom = Math.random
+
+function withSafeTimers(fn: () => void) {
+  const currentSetTimeout = globalThis.setTimeout
+  const currentRandom = globalThis.Math.random
+
+  try {
+    globalThis.setTimeout = safeSetTimeout
+    globalThis.Math.random = safeRandom
+
+    const result = fn()
+    return result
+  }
+  finally {
+    globalThis.setTimeout = currentSetTimeout
+    globalThis.Math.random = currentRandom
+  }
+}
 
 export const rpc = () => {
   const { rpc } = getWorkerState()
